Fixes issue with "temp errors"
This branch stops the false error message that were being produce from occurring and being sent to SERIAL_PORT. See https://github.com/MarlinFirmware/Marlin/issues/19994#issue-734104840.

Added a new ability that when the Adafruit MAX31865 board is checked for faults, the code will now display which of the seven (7) possible faults it found on the SERIAL_PORT with an appropriate error message.  Before it just displayed "Error: temp measurement error MAX6675", which does not indicate the fault condition of the Adafruit MAX31865.  After a fault is found and message is displayed, the code will clear the fault on the board.

Fixed the incorrect temperature being displayed by the Adafruit MAX31865 board connected to TEMP_SENSOR_1.
This error was caused by the fact that Marlin never instantiated and object for the second Adafruit MAX31865 which meant it was never initialized and the raw data being produced by the board was never getting converted because Marlin did not handle the data coming from the board and convert it to Celsius.  The wrong temperature value was read ~2047 at room temperature.

Marlin did display the correct temperature for the first Adafruit MAX31865 connected to TEMP_SENSOR_0.

Added a new ability that allows the user to use a (PT100 on E0 and a PT1000 on E1) OR (PT1000 on E0 and a PT100 on E1).  To do this I updated the MAX31865_SENSOR_OHMS to MAX31865_SENSOR_OHMS_0.  I also updated  MAX31865_CALIBRATION_OHMS to  MAX31865_CALIBRATION_OHMS_0.  I then added two additional Marlin variables which are used for TEMP_SENSOR_1 called MAX31865_SENSOR_OHMS_1 and MAX31865_CALIBRATION_OHMS_1.  With these four Marlin variables, the user can now use a PT100 and a PT1000 on the printer at the same time.

Added a check in SanityCheck.h that check to ensure that MAX31865_SENSOR_OHMS_1 and MAX31865_CALIBRATION_OHMS_1 are set if TEMP_SENSOR_1 is using the MAX31865 board.

See https://github.com/MarlinFirmware/Marlin/issues/19994#issue-734104840.  This branch fixes the "temp errors" being sent to the SERIAL_PORT.

Feature request fullfilled: https://github.com/MarlinFirmware/Marlin/issues/19663#issue-717855237